solana_tx_decoding

Function process_raydium_cpmm_swap_instruction

pub fn process_raydium_cpmm_swap_instruction(
    instruction: &Instruction<'_>,
    transfers: &[Instruction<'_>],
    running_token_balances: &mut HashMap<u8, u64>,
    block_time: u64,
    slot: u64,
    index: u64,
    atomic_instruction_index: u8,
    signers: &HashSet<Pubkey>,
    signature: &Signature,
) -> SwapTx

Process a Raydium Cpmm swap instruction and create a SwapTx. Assumes the instruction has been validated as a valid Raydium Cpmm swap. Uses token transfer instructions that follow the swap to determine swap amounts.